handled all the possible errors (backend)

+++ server/controllers/favoriteController.js
@@ -1,5 +1,7 @@
 const FavoriteModel = require('../models/favoriteModel.js')
 
+const AppError = require('../utlis/appError.js')
+
 
 
 
@@ -9,17 +11,30 @@ Description:  Fetch all the  favorites
 
 Method: GET
 
-Route: /api/favorites
+Route: /api/watch-later
 
 Access: Private
 
 */
 
-const fetchFavorites = async (req, res) => {
+const fetchFavorites = async (req, res, next) => {
 
-    const favorites = await FavoriteModel.find({}).sort({ createdAt: -1 })
+    try {
 
-    res.status(200).json(favorites)
+        const favorites = await FavoriteModel.find({}).sort({ createdAt: -1 })
+
+        res.status(200).json(favorites)
+
+    }
+
+
+    catch(error) {
+
+        return next(new AppError(error.message, 404, error))
+
+    }
+
+
 
 }
 
@@ -40,22 +55,35 @@ Access: Private
 */
 
 
-const searchFavorite = async (req, res) => {
-
-    const { id } = req.params
+const searchFavorite = async (req, res, next) => {
 
 
-    // checking whether any favorite exits with that imdbID or not
-    const favorite = await FavoriteModel.findOne({ imdbID: id })
+    try {
 
-    
-    if (!favorite) {
+        const { id } = req.params
 
-        return res.status(400).json({ "msg": "No favorite exits with that imdbID" })
+        // checking whether any favorite exits with that imdbID or not
+        const favorite = await FavoriteModel.findOne({ imdbID: id })
+
+        
+        if (!favorite) {
+
+        return next(new AppError('No favorite exits with that imdbID', 400))
+            
+        }
+
+
+      res.status(200).json(favorite)
+        
     }
 
+   
+    catch (error) {
 
-    res.status(200).json(favorite)
+        return next(new AppError(error.message, 404, error))
+
+    }
+    
 
 }
 
@@ -76,30 +104,21 @@ Access: Private
 
 */
 
-const createFavorite = async (req, res) => {
-
-    const { imdbID } = req.body
+const createFavorite = async (req, res, next) => {
 
     try {
 
-        // trying to find if the movie already exits in the database or not
-        const alreadyExists = await FavoriteModel.findOne({ imdbID: imdbID })
+         /* In the following, I am using spread operator to create  a object in the database with whatever there is in the 'req.body'. But actually, I can't simply put whatever I want in the req.body. Only some properties are accepted and they are defined in the mongoose schema. So, I need to make sure in the fronted that I only send those properties in the req.body.  */
 
-        if (alreadyExists) {
-            return res.status(404).json({ "msg": "The movie already exits in the database." })
-        }
-
-        
-        /* if the movie doesn't already exits in the database, we will add it to the database.
-        
-         In the following, I am using spread operator to create  a object in the database with whatever there is in the 'req.body'. But actually, I can't simply put whatever I want in the req.body. Only some properties are accepted and they are defined in the mongoose schema. So, I need to make sure in the fronted that I only send those properties in the req.body.  */
         const favorite = await FavoriteModel.create({ ...req.body })
 
         res.status(200).json(favorite)
     }
 
     catch (error) {
-        res.status(400).json({ error: error.message })
+
+       return next(new AppError(error.message, 404, error))
+
     }
 
 }
@@ -121,26 +140,36 @@ Access: Private
 */
 
 
-const deleteFavorite = async (req, res) => {
+const deleteFavorite = async (req, res, next) => {
 
-    const { id } = req.params
+    try {
+
+        const { id } = req.params
+
+        // trying to find the favorite 
+        const existsOrNot = await FavoriteModel.findOne({ imdbID: id })
 
 
+        if (!existsOrNot) {
 
-    // trying to find the favorite and delete it
-    const favorite = await FavoriteModel.findOneAndDelete({ imdbID: id })
+            return next (new AppError('There is no favorite with that imdbID.', 400))
+            
+        }
+
+        const favorite = await FavoriteModel.findOneAndDelete({ imdbID: id })
 
 
-    if (!favorite) {
-
-        return res.status(404).json({ "msg": "There is no favorite with that id." })
+        res.status(200).json(favorite)
     }
 
 
-    res.status(200).json(favorite)
+    catch (error) {
+
+        return next(new AppError(error.message, 404, error))
+ 
+     }
 
 }
-
